feat: introduce Le Chez brand identity and clean up imports (#11)
* feat: introduce Le Chez brand identity and clean up unused imports

- Add LeChezBrandAgent for documentation with smoking cheese personality
- Create comprehensive brand guide capturing early 2000s Flash animation energy
- Integrate Le Chez mascot into README with existential tagline
- Remove unused imports across all modules for cleaner codebase
- Minor code improvements including set operations and variable cleanup

* chore: remove brand file

* fix: add explicit toDouble conversion to avoid deprecated implicit Long->Double

- Added comment explaining why explicit conversion is needed
- ujson.Num expects Double parameter, explicit conversion avoids precision loss warning

* fix: resolve compilation errors in test files

- Revert underscore parameters in method definitions (not allowed in Scala)
- Keep underscore for unused lambda parameters (routes) where allowed
- Note: validatedRequest params in TestUserCrudAPI remain unused but are required by CaskChez trait

+++ CaskChez/src/main/scala/caskchez/Validation.scala
@@ -334,7 +334,8 @@ object ValidationHelpers {
       // Try parsing as number
       Try(value.toDouble) match {
         case Success(num) if value.contains(".") => ujson.Num(num)
-        case Success(num) if num.isWhole => ujson.Num(num.toLong)
+        // Explicit .toDouble needed as ujson.Num expects Double; avoids deprecated implicit Long->Double conversion
+        case Success(num) if num.isWhole => ujson.Num(num.toLong.toDouble)
         case Success(num) => ujson.Num(num)
         case Failure(_) => ujson.Str(value)
       }
